minor cosmetic changes

+++ db/migrate/20181011180116_create_readings.rb
@@ -2,7 +2,7 @@ class CreateReadings < ActiveRecord::Migration[5.2]
   def up
     create_table :readings do |t|
       t.references :thermostat, foreign_key: true
-      t.float :temperature
+      t.integer :temperature
       t.integer :battery_charge
       t.integer :humidity
       t.integer :reading_id
